added stock controller, api setup

+++ AuthService/AuthService/DTOs/StockDto.cs
@@ -6,7 +6,7 @@
         public int Id { get; init; }
         public string Ticker { get; init; } = "";
         public string DateOfPrediction { get; init; } = ""; 
-        public float PredictedReturn { get; init; }
+        public double PredictedReturn { get; init; }
         public int DaysPredicted { get; init; } 
     }
 }
