+++ AuthService/AuthService/Entities/Stock.cs
@@ -6,7 +6,7 @@
         public int Id { get; set; }
         public string Ticker { get; set; } = "";
         public string DateOfPrediction { get; set; } = "";  
-        public float PredictedReturn { get; set; }
+        public double PredictedReturn { get; set; }
         public int DaysPredicted { get; set; } 
     }
 }
